+++ src/imio/actionspanel/utils.py
@@ -1,12 +1,7 @@
 import logging
 logger = logging.getLogger('imio.actionspanel')
 
-from AccessControl.SecurityManagement import newSecurityManager
-from AccessControl.SecurityManagement import getSecurityManager
-from AccessControl.SecurityManagement import setSecurityManager
-
-from Products.CMFCore.utils import getToolByName
-from Products.CMFCore.tests.base.security import OmnipotentUser
+from plone import api
 
 
 def unrestrictedRemoveGivenObject(object_to_delete):
@@ -18,33 +13,13 @@ def unrestrictedRemoveGivenObject(object_to_delete):
       to handle.  This is called by the 'remove_givenuid' view that does the checks if user
       has at least the 'Delete objects' permission on the p_object_to_delete.
     """
-    # save current SecurityManager to fall back to it after deletion
-    oldsm = getSecurityManager()
-    # login as an omnipotent user
-    portal = getToolByName(object_to_delete, 'portal_url').getPortalObject()
-    newSecurityManager(None, APOmnipotentUser().__of__(portal.aq_inner.aq_parent.acl_users))
     # removes the object
     parent = object_to_delete.aq_inner.aq_parent
-    logMsg = '%s at %s deleted by "%s"' % \
-             (object_to_delete.meta_type, object_to_delete.absolute_url_path(), oldsm.getUser().getId())
-    try:
+    logMsg = '{} at {} deleted by "{}"'.format(
+        object_to_delete.meta_type,
+        object_to_delete.absolute_url_path(),
+        api.user.get_current().getId()
+    )
+    with api.env.adopt_roles(['Manager']):
         parent.manage_delObjects(object_to_delete.getId())
         logger.info(logMsg)
-    except Exception, exc:
-        # in case something wrong happen, make sure we fall back to original user
-        setSecurityManager(oldsm)
-        raise exc
-    # fall back to original user
-    setSecurityManager(oldsm)
-
-
-class APOmnipotentUser(OmnipotentUser):
-    """
-      Omnipotent for imio.actionspanel.  Heritates from Products.CMFCore's OmnipotentUser
-      but add a missing 'has_role' method...
-    """
-    def has_role(self, roles, obj=None):
-        return True
-
-    def getGroups(self):
-        return []