+++ CHANGES.rst
@@ -4,7 +4,9 @@ Changelog
 1.27 (unreleased)
 -----------------
 
-- Nothing changed yet.
+- Use api.adopt_roles rather than create a super user to execute a "own" delete
+  action.
+  [sdelcourt]
 
 
 1.26 (2017-04-13)